requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
django-countries==7.6.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
